Funciona, corregir la precision de las operaciones

+++ resultmodel.h
@@ -14,15 +14,15 @@ class ResultModel : public QAbstractTableModel
     public:
         vector<QString> words;
         //vector<long int> ids;
-        vector<long long int> ids;
-        vector<long long int> similaridad;
+        vector<long int> ids;
+        vector<double> similaridad;
         ResultModel(QObject * parent);
         int rowCount(const QModelIndex & parent = QModelIndex()) const;
         int columnCount(const QModelIndex & parent = QModelIndex()) const;
         QVariant data(const QModelIndex & index, int role) const;
         void setWords(vector<QString> & v);
-        void setIds(vector<long long int> & v);
-        void setSimilaridad(vector<long long int> & v);
+        void setIds(vector<long int> & v);
+        void setSimilaridad(vector<double> & v);
 };
 
 #endif // RESULTMODEL_H
